refactor: move ConfigTypeFilterDropdown to it's own file
